+++ src/Pim/Bundle/InstallerBundle/Command/InstallCommand.php
@@ -8,6 +8,7 @@ use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 use Pim\Bundle\InstallerBundle\CommandExecutor;
+use Symfony\Component\Filesystem\Filesystem;
 
 /**
  * Override OroInstaller command to add PIM custom rules
@@ -23,11 +24,6 @@ class InstallCommand extends ContainerAwareCommand
      */
     const APP_NAME = 'Akeneo PIM';
 
-    const TASK_ALL    = 'all';
-    const TASK_ASSETS = 'assets';
-    const TASK_CHECK  = 'check';
-    const TASK_DB     = 'db';
-
     /**
      * @var CommandExecutor $commandExecutor
      */
@@ -41,14 +37,7 @@ class InstallCommand extends ContainerAwareCommand
         $this
             ->setName('pim:install')
             ->setDescription(sprintf('%s Application Installer.', static::APP_NAME))
-            ->addOption('force', null, InputOption::VALUE_NONE, 'Force installation')
-            ->addOption(
-                'task',
-                null,
-                InputOption::VALUE_REQUIRED,
-                'Determines tasks called for installation (can be all, check, db or assets)',
-                self::TASK_ALL
-            );
+            ->addOption('force', null, InputOption::VALUE_NONE, 'Force installation');
     }
 
     /**
@@ -83,22 +72,14 @@ class InstallCommand extends ContainerAwareCommand
         $output->writeln('');
 
         try {
-            switch ($input->getOption('task')) {
-                case self::TASK_CHECK:
-                    $this->checkStep($input, $output);
-                    break;
-                case self::TASK_DB:
-                    $this->databaseStep($input, $output);
-                    break;
-                case self::TASK_ASSETS:
-                    $this->assetsStep($input, $output);
-                    break;
-                default:
-                    $this->checkStep($input, $output)
-                        ->databaseStep($input, $output)
-                        ->assetsStep($input, $output);
-                    break;
-            }
+
+            $this->cleanDirectory($this->getContainer()->getParameter('upload_dir'));
+            $this->cleanDirectory($this->getContainer()->getParameter('archive_dir'));
+
+            $this
+                ->checkStep($input, $output)
+                ->databaseStep($input, $output)
+                ->assetsStep($input, $output);
         } catch (\Exception $e) {
             return $e->getCode();
         }
@@ -175,5 +156,19 @@ class InstallCommand extends ContainerAwareCommand
         $params = $dumper->parse();
         $params['system']['installed'] = $installed;
         $dumper->dump($params);
+    }
+
+    /**
+     * Remove directory and all subcontent
+     *
+     * @param string $directory
+     */
+    protected function cleanDirectory($folder)
+    {
+        $filesystem = new Filesystem();
+        if ($filesystem->exists($folder)) {
+            $filesystem->remove($folder);
+        }
+        $filesystem->mkdir($folder);
     }
 }